Move blog to posts dir, fix MDX
