+++ gatsby-node.js
@@ -1,80 +1,69 @@
-const _ = require('lodash')
 const path = require('path')
 const { createFilePath } = require('gatsby-source-filesystem')
 
-exports.createPages = ({ graphql, actions }) => {
+exports.createPages = async ({ graphql, actions, reporter }) => {
   const { createPage } = actions
+  const blogSingleTemplate = path.resolve('./src/templates/blog-single.js')
 
-  return new Promise((resolve, reject) => {
-    const blogSingle = path.resolve('./src/templates/blog-single.js')
-    const projectSingle = path.resolve('./src/templates/project-single.js')
-
-    resolve(
-      graphql(
-        `
-          {
-            allMdx(
-              sort: { fields: [frontmatter___date], order: DESC }
-              limit: 1000
-            ) {
-              edges {
-                node {
-                  fields {
-                    slug
-                  }
-                  frontmatter {
-                    title
-                  }
-                }
+  const result = await graphql(
+    `
+      query {
+        allMdx(
+          sort: { fields: [frontmatter___date], order: DESC }
+          limit: 1000
+        ) {
+          edges {
+            node {
+              fields {
+                slug
+              }
+              frontmatter {
+                title
               }
             }
           }
-        `
-      ).then((result) => {
-        if (result.errors) {
-          console.log(result.errors)
-          reject(result.errors)
         }
+      }
+    `
+  )
 
-        // Create blog posts pages.
-        const posts = [...result.data.allMdx.edges]
+  if (result.errors) {
+    reporter.panicOnBuild(`Error while running GraphQL query.`)
+    return
+  }
 
-        _.each(posts, (post, index) => {
-          const previous =
-            index === posts.length - 1 ? null : posts[index + 1].node
-          const next = index === 0 ? null : posts[index - 1].node
+  // Create blog posts pages.
+  const posts = [...result.data.allMdx.edges]
 
-          const path = post.node.fields.slug
-          let component = blogSingle
+  posts.forEach((post, index) => {
+    const slug = post.node.fields.slug
+    let component = blogSingleTemplate
 
-          // skip projects
-          if (path.startsWith('/project')) {
-            return null
-          }
-          createPage({
-            path,
-            component,
-            context: {
-              slug: post.node.fields.slug,
-              previous,
-              next,
-            },
-          })
-        })
-      })
-    )
+    // skip projects
+    if (slug.startsWith('/project')) {
+      return
+    }
+
+    createPage({
+      path: slug,
+      component,
+      context: {
+        slug,
+      },
+    })
   })
 }
 
 exports.onCreateNode = ({ node, actions, getNode }) => {
   const { createNodeField } = actions
 
-  if (`allMdx JavascriptFrontmatter`.includes(node.internal.type)) {
-    const value = createFilePath({ node, getNode })
+  if (node.internal.type === 'Mdx') {
+    let slug = createFilePath({ node, getNode })
+    slug = '/blog' + slug
     createNodeField({
-      name: `slug`,
       node,
-      value,
+      name: `slug`,
+      value: slug,
     })
 
     // https://github.com/kentcdodds/kentcdodds.com/blob/9149d6bde95e1694f8957cf978ba58faa2b27d9b/gatsby-node.js#L404-L411
