+++ src/posts/2017/creating-static-websites-with-react-part-1.md
